pyproject.toml
[tool.poetry]
name = "plaster_yaml"
version = "0.2.0"
description = "A plaster plugin to configure pyramid app with Yaml"
readme = "README.md"
authors = ["Guillaume Gauvrit <guillaume@gauvr.it>"]
homepage = "https://github.com/mardiros/plaster-yaml"
repository = "https://github.com/mardiros/plaster-yaml"
include = ["CHANGELOG.md"]
license = "BSD-derived"

[tool.poetry.dependencies]
# See: https://pypi.org/project/backports.entry-points-selectable
importlib_metadata = { version = ">=3.6", python = "<3.10" }
python = "^3.7"
plaster = "^1.0"
# Upgraded PyYAML from ^5.4 due to: https://github.com/yaml/pyyaml/issues/601
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
black = "^21.4b0"
isort = "^5.8.0"
# Tests
gunicorn = "^20.1.0"
PasteDeploy = "^2.1.1"
pytest = "^5.2"
waitress = "^2.0.0"
pyramid-helloworld = "^0.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
